tech_products/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
graphviz==0.20.1
mysqlclient==2.2.0
pydot==1.4.2
PyJWT==2.8.0
pyparsing==3.1.1
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
